requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
boto3==1.28.85
